[Android] Use strings instead of ints in java preferences code
Chromium change:
https://chromium.googlesource.com/chromium/src/+/d3ea46913e4f61fbd84b4b71dbde97feb8f1e625

Use strings instead of ints in java preferences code.

This will allow us to remove the centralized declaration of prefs
in //chrome, later permitting componentization of PrefServiceBridge.

PrefServiceBridge.java and PrefRegistrar.java are updated to use the
preference string keys from C++ instead of an integer that is mapped
to the string in C++.

This uses java_cpp_strings with the relevant prefs files to generate
the java string definitions. These pref files include both //chrome
and //components prefs. Later, some or all of the //components prefs
can be moved to a different java_cpp_strings target.

Bug: 1071603