Avoid DCHECK in BrowserProcessImpl::GetApplicationLocale.
ChromeComponentExtensionResourceManager::Data::Data adds PDF strings by
calling pdf_extension_util::AddStrings, which, in turn calls into
BrowserProcessImpl::GetApplicationLocale that has a DCHECK to make sure
the method is only called on the UI thread.

In Chromium that is true in normal operation, because
ChromeComponentExtensionResourceManager::Data gets LazyInit'ed early
when the CryptoTokenExtension is loaded on the UI thread.

We don't load that extension, so the first time LazyInit is called
happes to be on a worker thread, resulting in the DCHECK.

Added code to BraveExtensionsBrowserClientImpl to force LazyInit on the
UI thread after creation of the ChromeComponentExtensionResourceManager
instance.

Chromium change:

https://chromium.googlesource.com/chromium/src/+/3556ea28e8ae03eeb8ee203675fd73bf81b88527

commit 3556ea28e8ae03eeb8ee203675fd73bf81b88527
Author: dpapad <dpapad@chromium.org>
Date:   Tue Jun 16 20:56:41 2020 +0000

    PDF Viewer: Use $i18n{} to populate |lang| and |dir| HTML attributes.

    Bug: 1091516